feat(frontend): add step to UI flow with paragraph reading
