+++ frontend/src/routes/+page.svelte
@@ -4,21 +4,6 @@
     import { goto } from '$app/navigation';
 
     let formData = { context: '', userAnswer: '', correctAnswer: '', question: '' };
-    const minimumAnswerLength = 5;
-
-    const handleSubmit = () => {
-        const answerInput = document.querySelector('.answer-input');
-        const answer = String(answerInput.value);
-        if (answer.length < minimumAnswerLength) {
-            alert(`Answer is too short. Minimum length: ${minimumAnswerLength}.`)
-            return;
-        }
-
-        if (typeof window !== 'undefined') {
-            sessionStorage.setItem('formData', JSON.stringify(formData));
-            goto('/evaluate');
-        }
-  };
 
     /**
 	 * @type {any[]}
@@ -72,16 +57,7 @@
         question = result.data.question;
     }
 
-    /**
-     * Make Ctrl + Enter send a form.
-	 * @param {{ key: any; }} event
-	 */
-    function handleKeyDown(event) {
-        // @ts-ignore
-        if (event.ctrlKey && event.key === 'Enter') {
-            handleSubmit();
-        }
-  }
+
 
 </script>
 
@@ -98,45 +74,13 @@
         </div>
     {/each}
 {:else}
-<form on:submit|preventDefault={handleSubmit}>
-    <p>
-        <b>Question:</b>
-        <br>
-        {question}
-    </p>
-    <input type="text" bind:value={formData.question} hidden readonly/>
-    <input type="text" bind:value={formData.context} hidden readonly/>
-    <input type="text" bind:value={formData.correctAnswer} hidden readonly/>
-    <label>
-        <p><b>Your answer:</b></p>
-        <textarea
-            class="answer-input"
-            bind:value={formData.userAnswer}
-            on:keydown={handleKeyDown}
-            required
-            autofocus
-        ></textarea>
-    </label>
-    <button type="submit">&rarr;</button>
-    </form>
+    <p>Read a paragraph</p>
+    <p>{formData.context}</p>
+    <button on:click={() => goto('/answer')}>&rarr;</button>
 {/if}
 
 
 <style>
-    label > p {
-        /* padding: 10px; */
-        display: block;
-    }
-
-    label {
-        height: 10vh;
-    }
-
-    .answer-input {
-        width: 80%;
-        height: 10rem;
-    }
-
     .learning-material {
         margin: 50px 0;
         display: flex;
